requirements.txt
flask>=3.0
flask-cors>=4.0
groq>=0.11
python-dotenv>=1.0
